ASCII-Artify/requirements.txt
pillow==11.3.0
rich==14.1.0
